ai/yeopseo/requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
dill==0.2.8.2
dominate==2.7.0
entrypoints==0.4
fastjsonschema==2.16.2
idna==3.4
imageio==2.22.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
lxml==4.9.1
MarkupSafe==2.1.1
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.7
numpy==1.23.3
packaging==21.3
pandocfilters==1.5.0
Pillow==9.2.0
Pygments==2.13.0
pyparsing==3.0.9
pyreadline==2.1
pyrsistent==0.18.1
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==304
pyzmq==24.0.1
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
soupsieve==2.3.2.post1
tifffile==2022.8.12
tinycss2==1.1.1
torch==1.12.0
torchvision==0.13.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
